requirements.txt
PySide6
google-genai
python-telegram-bot
psutil
send2trash
